# Performance improvement with cell reuse