--- a/.travis.yml
+++ b/.travis.yml
@@ -1,8 +1,8 @@
 language: swift
-osx_image: xcode9.4
+osx_image: xcode10
 
 script: 
-  xcodebuild -workspace CollectionKit.xcworkspace -scheme CollectionKit -sdk iphonesimulator -destination 'platform=iOS Simulator,name=iPhone 7,OS=10.3.1' build test
+  xcodebuild -workspace CollectionKit.xcworkspace -scheme CollectionKit -sdk iphonesimulator -destination 'platform=iOS Simulator,name=iPhone XR,OS=12.0' build test
 
 after_success:
   - bash <(curl -s https://codecov.io/bash) -J 'CollectionKit'