--- a/.travis.yml
+++ b/.travis.yml
@@ -4,5 +4,8 @@ osx_image: xcode10
 script: 
   xcodebuild -workspace CollectionKit.xcworkspace -scheme CollectionKit -sdk iphonesimulator -destination 'platform=iOS Simulator,name=iPhone XR,OS=12.0' build test
 
+before_install:
+  - pod repo update
+
 after_success:
   - bash <(curl -s https://codecov.io/bash) -J 'CollectionKit'