--- a/Podfile
+++ b/Podfile
@@ -3,6 +3,6 @@ use_frameworks!
 
 target "CollectionKitTests" do
 
-	pod 'Quick', "= 1.3.0"
-	pod 'Nimble', "= 7.1.1"
+	pod 'Quick', "= 1.3.2"
+	pod 'Nimble', "= 7.3.1"
 end