… .swift-version
… Podfile.lock