Regarding https://github.com/SoySauceLab/CollectionKit/issues/97
@Banck gives us a brilliant idea of not removing the cells from CollectionView.
This brings major improvement in main thread cpu time.

With 10s scrolling session, reuseManager's `lifeSpan` set to 5.0s. and scrolling from top to the end every 1 second. We get the following result on an iPhone X.

Notices that `CollectionView.loadCells()` cpu time changed from 558ms to 288ms.

### Before

<img width="1025" alt="screen shot 2018-09-22 at 9 59 41 pm" src="https://user-images.githubusercontent.com/3359850/45928978-780fb280-bf00-11e8-8b1e-841f0ed7db75.png">


### After

<img width="995" alt="screen shot 2018-09-23 at 7 41 03 am" src="https://user-images.githubusercontent.com/3359850/45929366-75638c00-bf05-11e8-8f51-0dd0f09ee721.png">


### UICollectionView
And just for comparison with UICollectionView.
UICollectionView has about the same scroll performance as CollectionKit after this patch.
<img width="840" alt="screen shot 2018-09-23 at 7 53 35 am" src="https://user-images.githubusercontent.com/3359850/45929400-e7d46c00-bf05-11e8-8d6c-02470168395c.png">